Refactor network service types and update imports for better organization and avoid circular dependency issue
