+++ src/app/drive/services/network.service/types/index.ts
@@ -1,0 +1,16 @@
+export interface IUploadParams {
+  filesize: number;
+  filecontent: File;
+  isUploadedFromFolder?: boolean;
+  progressCallback: ProgressCallback;
+}
+
+export type ProgressCallback = (progress: number, uploadedBytes: number | null, totalBytes: number | null) => void;
+
+export interface EnvironmentConfig {
+  bridgeUser: string;
+  bridgePass: string;
+  encryptionKey: string;
+  bucketId: string;
+  useProxy: boolean;
+}
